cffi build: check if link already created.

+++ cffi_build/cffi_build.py
@@ -726,8 +726,9 @@ ffibuilder.set_source(
 # For some reason, cffi makes it such that it looks for libv_rep.so.1
 # rather than libv_rep.so. So we add a symlink.
 path = os.path.join(os.environ['VREP_ROOT'], 'libv_rep.so')
-print('creating symlink: %s -> %s' % (path + '.1', path))
-os.symlink(path, path + '.1')
+if not os.path.exists(path + '.1'):
+    print('creating symlink: %s -> %s' % (path + '.1', path))
+    os.symlink(path, path + '.1')
 
 if __name__ == "__main__":
     ffibuilder.compile(verbose=True)
